refactor modal component
