+++ src/components/modal/index.tsx
@@ -1,2 +1,4 @@
 export {Modal} from './Modal';
 export type {IModalProps} from './Modal';
+export type {IHeaderComponentProps} from './ModalHeader';
+export type {IFooterComponentProps} from './ModalFooter';
